refactor: move Lightbox to own file and created helper
