+++ src/components/Gallery/helper.ts
@@ -1,0 +1,3 @@
+export const getImageNumber = (name: string): number => {
+  return Number(name.split('.')[0]);
+};
