+++ src/components/Lightbox/index.tsx
@@ -1,0 +1,202 @@
+import Image from 'next/image';
+import React, { useEffect, useCallback } from 'react';
+
+interface GalleryImage {
+  id: string;
+  url: string;
+  name: string;
+}
+
+interface LightboxProps {
+  images: GalleryImage[];
+  currentIndex: number;
+  isOpen: boolean;
+  onClose: () => void;
+  onNext: () => void;
+  onPrevious: () => void;
+}
+
+export default function Lightbox({
+  images,
+  currentIndex,
+  isOpen,
+  onClose,
+  onNext,
+  onPrevious,
+}: LightboxProps) {
+  const handleKeyDown = useCallback(
+    (event: KeyboardEvent) => {
+      if (!isOpen) return;
+
+      switch (event.key) {
+        case 'Escape':
+          onClose();
+          break;
+        case 'ArrowLeft':
+          onPrevious();
+          break;
+        case 'ArrowRight':
+          onNext();
+          break;
+      }
+    },
+    [isOpen, onClose, onNext, onPrevious]
+  );
+
+  useEffect(() => {
+    document.addEventListener('keydown', handleKeyDown);
+
+    // Prevent body scroll when lightbox is open
+    if (isOpen) {
+      document.body.style.overflow = 'hidden';
+    } else {
+      document.body.style.overflow = 'auto';
+    }
+
+    return () => {
+      document.removeEventListener('keydown', handleKeyDown);
+      document.body.style.overflow = 'auto';
+    };
+  }, [handleKeyDown, isOpen]);
+
+  if (!isOpen || !images[currentIndex]) return null;
+
+  const currentImage = images[currentIndex];
+
+  return (
+    <div
+      style={{
+        position: 'fixed',
+        top: 0,
+        left: 0,
+        width: '100%',
+        height: '100%',
+        backgroundColor: 'rgba(0, 0, 0, 0.9)',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+        zIndex: 1000,
+      }}
+      onClick={onClose}
+    >
+      <button
+        onClick={onClose}
+        style={{
+          position: 'absolute',
+          top: '20px',
+          right: '20px',
+          background: 'rgba(255, 255, 255, 0.8)',
+          border: 'none',
+          borderRadius: '50%',
+          width: '40px',
+          height: '40px',
+          fontSize: '20px',
+          cursor: 'pointer',
+          zIndex: 1001,
+        }}
+        aria-label="Close lightbox"
+      >
+        ×
+      </button>
+
+      {/* Previous button */}
+      {images.length > 1 && (
+        <button
+          onClick={(e) => {
+            e.stopPropagation();
+            onPrevious();
+          }}
+          style={{
+            position: 'absolute',
+            left: '20px',
+            top: '50%',
+            transform: 'translateY(-50%)',
+            background: 'rgba(255, 255, 255, 0.8)',
+            border: 'none',
+            borderRadius: '50%',
+            width: '50px',
+            height: '50px',
+            fontSize: '20px',
+            cursor: 'pointer',
+            zIndex: 1001,
+          }}
+          aria-label="Previous image"
+        >
+          ‹
+        </button>
+      )}
+
+      {/* Next button */}
+      {images.length > 1 && (
+        <button
+          onClick={(e) => {
+            e.stopPropagation();
+            onNext();
+          }}
+          style={{
+            position: 'absolute',
+            right: '20px',
+            top: '50%',
+            transform: 'translateY(-50%)',
+            background: 'rgba(255, 255, 255, 0.8)',
+            border: 'none',
+            borderRadius: '50%',
+            width: '50px',
+            height: '50px',
+            fontSize: '20px',
+            cursor: 'pointer',
+            zIndex: 1001,
+          }}
+          aria-label="Next image"
+        >
+          ›
+        </button>
+      )}
+
+      {/* Image container */}
+      <div
+        style={{
+          maxWidth: '90%',
+          maxHeight: '90%',
+          position: 'relative',
+        }}
+        onClick={(e) => e.stopPropagation()}
+      >
+        <Image
+          src={currentImage.url}
+          alt={currentImage.name}
+          width={1200}
+          height={800}
+          style={{
+            maxWidth: '100%',
+            maxHeight: '90vh',
+            width: 'auto',
+            height: 'auto',
+            objectFit: 'contain',
+          }}
+          unoptimized
+        />
+
+        {/* Image info */}
+        <div
+          style={{
+            position: 'absolute',
+            bottom: '-40px',
+            left: '0',
+            right: '0',
+            textAlign: 'center',
+            color: 'white',
+            fontSize: '14px',
+          }}
+        >
+          <p>{currentImage.name}</p>
+          <p>
+            {currentIndex + 1} of {images.length}
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export type { LightboxProps, GalleryImage };
